Modified SGE spec and worked more on Pygame implementation.
I started work on the Sprite class in the Pygame implementation. When I was
doing so, I noticed that the Sprite class specification was inconsistent with
everything else, so I fixed this inconsistency.

Also added some of the code I wrote for the Pygame implementation to the
template, since it is entirely centered around choosing the right file to
load and can be useful for other implementations.