Tweaked the specification to allow multiple views in rooms.
When I started implementing views in the Pygame implementation, I
noticed how odd it was that views supported positioning when only
one view could be in a given room. At first I considered getting
rid of this, but I decided that the ability to have multiple views,
all active at once, could be a very useful feature, usable for easy
split-screening, so I changed the specification to allow multiple
views at once.

I also noticed that some parts of the documentation were worded in
such a way as to suggest that certain variables could return IDs,
when I only meant that they could be *assigned* to IDs. This has
been clarified. In the process, some of these have been made
read-only in the specification since it makes no sense for them
to be modifiable (they were ``layers`` of Background and ``objects``
of Room).