Added repositioning of _PygameSprite for the easy case.
That is, the case where there is only one view that takes up the
entire screen; here, all that's needed is to change the rect. Every
other case is next, which will need to be much more complicated
since you have cases of the sprite being at the edge of the view
and the object being visible on more than one view.