Changed sprites to use a blank image by default, not a black image.
This is the first step toward what I call "dynamic sprites", which
is intended to replace the Game Maker-like drawing system. I think
Game Maker's drawing system is counter-intuitive for beginners and
it is extremely difficult to debug in the Pygame implementation of
SGE.

Also, failing to find a specified image now raises an exception,
allowing game programmers to decide what to do in those case. I
plan to similarly change the loading of other resources later.