Fixed SGE using the current working directory and not the program directory.
The fix was simple: I added sys.argv[0] to the start of all the
default paths. This doesn't help if the program was located
implicitly by the OS (i.e. it was located in $PATH after typing
just the program name), but the location of data is best changed
in those cases anyway, so it should be fine.

I recommend for any script which changes the data directories to
not use this method, but to instead put ``os.path.dirname(__file__)``
as the base directory. For example:

image_directories.append(os.path.join(os.path.dirname(__file__), 'images'))