Removed the requirement for implementations to support certain formats.
I previously made it a requirement for implementations to support
Ogg Vorbis and uncompressed WAV for sounds, and Ogg Vorbis for
music. I noticed that there was not such a requirement for images
and, thinking it over, decided that the requirement doesn't make
much sense. Instead, what formats are supported is up to each
individual implementation, and I will add a suggestion of common
formats to use in the API reference for both images and sounds.