Changed the structure of the documentation.
I got rid of the ".txt" extension from every file, since it was
creating an inconsistency and is. COPYING and COPYING.LESSER have
been removed from the base directory, since they don't actually
apply to everything and they are already included where they do
apply (in sge-pygame and sge-template). LICENSES.txt has been
replaced with a LICENSES file in every directory that has non-functional
data files, each LICENSES file describing the licenses only for
the exact directory it is in.