Changed bbox_x and bbox_y of Sprite to default to the top-left of the image.
Previously, it defaulted to 0, which was origin_x or origin_y. That
was the top-left of the image by default, but if you changed the
origin, this would change along with it, and this is almost never a
desired behavior. These attributes now default to None, and when
set to None, they become -origin_x or -origin_y, which always results
in the bounding box starting at the top-left corner of the image
regardless of where the origin is.