Changed some methods in sge.Music to be static methods.
Having to stop the specific music object that was playing was
gratuitous; you can only have one music file playing at a time.
Also, it made it unclear that you are not supposed to be able to
pause a composition, play another, and then unpause the original;
it is more clear now that these methods are static.