Changed Background.[x,y] to Room.background_[x,y].
Making them attributes of the background was silly; a single
background will be typically shared by many rooms and can be
positioned differently in each of them, which was the whole point
of Background.x in the first place. Making it an attribute of Room
makes it easier to use for this purpose.