Added the source generator and room editor to the repo.
I've kept these to myself up to this point (except handing the source
generator to Emilio), but because I'm planning to distribute versions
of these with SGE itself anyway, I'm adding them to this repo. The
source generator may be a bit outdated and the room editor is
incomplete; this will be corrected over time.