Reverting to storing rooms in separate files.
The proposed format had everything in one file, but I think keeping
the rooms separate is a good idea for two reasons: one, it makes it
a lot easier to prevent the room editor from causing editing
conflicts with the main GUI. Two, I was going to change sourcegen.py
to store rooms in separate JSON files anyway. Putting them in
separate JSON files in the first place is more convenient.

The new proposed format basically moves everything that was going to
be in a "rooms" list in the main file into a bunch of separate JSON
files which are referenced by the main file much like how functions
are referenced.