Changed all references to "SGE" to say "the SGE".
I find that there's a strong temptation to say "es-jee-ee" when
"SGE" is presented as a proper noun, so I've changed all references
to the SGE in the documentation to say "the SGE" instead of just
"SGE". This makes sense, too, because it stands for "Stellar Game
Engine", and I've always said e.g. "The Stellar Game Engine is..."
and not just "Stellar Game Engine is...".

Also made some other small changes; for example, "SGE Pygame" has
been renamed to "the Pygame SGE", and the message shown when DEBUG
is true in the Pygame SGE is now "SGE Art of Debugging Enabled" (a
small Naruto reference).