Reworked the doc build structure.
Previously, the doc building structure was only suited to building
one version of the docs at a time. I moved all the stuff that
shouldn't change to a "common" directory and then gave a directory
specifically for the Pygame SGE to build its docs.

Since each implementation will have its own docs in the future, the
plan is for there to be a "doc" folder in sge-pygame.  There will
be no "doc" folder in the base directory in the next release, or
perhaps it will be a symbolic link to the docs of the Pygame SGE.