Added the STJ: Stellar JSON Library
This is a small module (though I worked on it for quite a while)
which makes it very easy to read and write the Stellar JSON format.
No need to manually edit the JSON files anymore. This will make
using the files a lot easier.

I did some basic testing and it seems to work without problems.