Fixed a problem with projections.
Stupid me, I used implicit arguments, and when the ID argument
became the second one rather than one of the last ones, that caused
a width assignment to be assigned as an ID. This screwed up the look
of anything projected and caused a crash if there were multiple
projections with the same attempted width assignment. This has been
fixed now.