Removed text in the docstrings that was too specific about dialogs.
Dialog boxes should be allowed to vary up to and including actual
toolkit libraries. I didn't use them in the Pygame SGE because it's
inconvenient to do so, but in another implementation it might be
easy, and yet these strict requirements would have made them
gratuitously no good as far as following the specification.

In addition to that, the controls specified previously don't
necessarily make sense for all devices.

Best to just leave this one to the implementations to decide.