Changed the license of stj from the GNU GPL to the GNU LGPL.
The reason for this is it, by default, generates code that depends
on it. The GNU GPL would require any game under a license other
than the GNU GPL to disable this option, and I don't want to
encourage this. I also certainly don't want the library to help
users violate its own license; that would be like a nasty trap.