Added a warning to scripts generated by the STJ that they aren't source.
"Source code" means the preferred form for editing, so in the case of
games developed with an IDE that stores the relevant data in a JSON
file, that JSON file and all related files are the source code, not
the generated Python file. This might not be obvious to everyone, so
a warning has been added mentioning this.