Changed the "loops" argument of sound playing methods.
Previously, I specified sge.Sound.play, sge.Music.play, and
sge.Music.queue to have a "loops" argument that works the way
Pygame describes its "loops" argument. It turns out that Pygame's
music "loops" argument is actually not implemented the way the docs
say, though the sound "loops" argument is, so I needed to fix this.
I decided to make it the way Pygame actually does it for music.