Fixed up the tools directory.
Deleted some obsolete code and data files, and fixed the LICENSES
files being wrongly placed.

I got rid of the version of the SELE that was made with the SGE
finally. This version just wasn't going anywhere. I can't see
myself going through with all the work necessary to make a SGE-
based interface work well, and I can't see anyone else doing it
either. It was a silly choice to begin with, and wxPython is a
much better choice even despite supporting only Python 2.

I might make a minimalist level editor based on keyboard shortcuts;
this will be much simpler to make in the SGE than a proper level
editor. But what I had already done in the SGE SELE isn't going to
be helpful for that.