Fixed a bug with resetting of rooms.
I neglected to *copy* a certain list, rather than just reassigning
it to another variable. This caused a crash any time a room ended
with a different set of objects than it started with and would have
otherwise screwed up the whole room resetting mechanic anyway. This
is fixed now.