Fixed and simplified directional collision detection.
The previous method for directional collision detection was naively
complicated and used xvelocity and yvelocity, a bad choice. I found
that setting xvelocity to 0 in a right collision could cause the
other object to assume that the opposite direction of collision was
occurring, which in my case teleported the character to the other
side of a wall. It's properly based on bounding boxes now.