Changed the way xprevious and yprevious work.
xprevious and yprevious previously reported the values of x and y
before they were last changed. The problem with this is you can't
know when this was; it could be last frame, earlier this frame, or
several frames ago, depending on whether or not the object was
moving. They now always report the values of x and y as they were
at the end of the previous frame, which is what is usually needed
from variables like these.