Added Sprite.speed and StellarClass.image_speed.
I realized when working on a project of mine that there are times
when adjusting the animation speed based on the FPS of the game
itself makes sense, for example to choose the animation speed of a
walking animation. So Sprite.speed and StellarClass.image_speed are
for that purpose; they are the animation rate as a factor of
sge.game.fps (the same as the "image_speed" variable in Game Maker).