Removed a bizarre arbitrary error.
There was code, when in debug mode, that arbitrarily caused an
error when the mouse went from invisible to visible. I'm sure I had
a reason for this, but I can't fathom what it was, and it breaks
the perfectly legitimate functionality of being able to make the
mouse visible.

Not making a note of this in WHATSNEW, because this hasn't actually
affected any distributions of the Pygame SGE; they all had debug
mode turned off.