Made Python 2 a second-class citizen.
I've been planning on doing this for a long time, and now I'm
finally doing it: the Pygame SGE, as it exists in the repo, has
been converted to Python 3. Python 2 versions will still be
distributed, but Python 2 will be a second-class citizen; the core
Python 3 code will be converted into Python 2 code.

Similarly, the __future__ imports have been removed from the spec,
and the examples have been converted to Python 3.

Also updated some copyright notices and added use of a custom mouse
sprite to the splitscreen example, since it occurred to me that I
hadn't tested that.