Changed the way room projections are done.
I was hoping this would make projections more efficient, but it
doesn't (at least not noticeably). However, I did fix a bug in the
process where sprites that were passed on to project_sprite would
be destroyed, and the new method is a bit better than the old
method: the old method kept the projection showing for an extra
frame to compensate for the fact that it couldn't tell for sure if
it was shown on the same frame it was called, while the new method
causes the projection to always show exactly once.