Fixed problems with going to fullscreen.
The only one that was visible was that the resolution wasn't being
set to the monitor's native resolution; rather, it was being set to
whatever the current size of the window was. Behind-the-scenes,
though, this was preventing an even worse bug where the screen was
not being updated outside an area of the previous window size. Both
of these problems have been fixed.