Fixed the bug with destroyed objects leaving artifacts behind.
Previously, this was hidden in the circle popper example because
the creation of a sprite every frame caused the "_background_changed"
attribute of sge.game True, which basically tells it to update the
entire screen. So to fix the problem with artifacts when objects
are destroyed, _background_changed has been set to True when
StellarClass.destroy is called.

Probably the root of the problem lies in the Pygame sprite being
killed too soon. I might investigate this later, but it's not very
important.