Fixed rotated precise collision detection not working.
This was a long-standing bug in the Pygame SGE that I was kind of
aware of, but put off because precise collision detection is so
ridiculously complicated in the first place. It basically works fine
now. It's not pixel-perfect precisely right if the window is being
scaled up (because it always happens at the same resolution), but
now it *works*, and that's what's really important.

Also, deleted object_moving.py. This was a completely useless
example; several other examples included a moving object plus other
things.