Partially re-wrote the collision system specification/implementation
I previously aimed to make the collision function the same as Game
Maker: if an object is rotated, it's about the origin, and collision
boxes are adjusted accordingly; also like this for scaling. But this
is a nightmare to implement; it adds several "offset" variables in
the background that have to be accounted for, and the result is a
horrible mess that is difficult to understand, let alone maintain.

So I've changed it. Now, objects can specify their origins, and this
is taken advantage of: by default, rotation is about the center of
the image, not the origin. There is a new variable for rotating
about the origin; it accomplishes this by adjusting the object's
origin on-demand. The resulting code is much, MUCH cleaner and
easier to read, understand, and maintain.

In addition, rectangle-based collision detection (including
ellipse collision) is not expected to scale with xscale and yscale,
or to rotate with rotation. In fact, they shouldn't, and the docs
explicitly mention this. I thought long and hard, and I couldn't
think of any legitimate reason one would need this handled for them
automatically where precise collision detection wouldn't be better.

In addition, this system might be better. One of the limitations of
the bounding boxes being automatically adjusted like that is if you
want to flip something, e.g. for a platformer, you have to make the
origin at the center. In addition, the origin not being in the exact
perfect position would cause a glitch where your bounding box would
suddenly change position when you move; this can cause you to get
stuck in a wall, for instance. No worries now that the bounding box
remains static.