Collision improvements.
This is a pretty huge one and it breaks compatibility, but it makes
the collision detection system much more versatile.

Firstly, the old StellarClass.collides method has been replaced
with StellarClass.collision, which is very similar, but rather than
just returning "True" or "False", it returns a list of colliding
objects. In the Pygame SGE, this is optimized; it doesn't check
anything outside of self's collision zone. It can now also accept
a list of objects, which is a potentially useful new feature.

Secondly, shape-based collision functions have been added to the new
collision module. There are now rectangle, ellipse, circle, and line
collision detection functions. This was a feature from Game Maker
which was missing in the Pygame SGE. Like with StellarClass.collides,
the Pygame SGE optimizes this collision detection; no unnecessary
checks are done.