Fixed origin being stuck at its initial value.
When image_origin_x and image_origin_y are None, they're supposed
to delegate to the sprite's value. I made the mistake of actually
setting the internal value to this value; this took away the
special None status, causing the origin to not automatically adjust
for a change of sprites. This has been fixed, and reading these
variables still transparently return the origin.