Clarified how Sprite.draw_sprite's positioning works.
I found that I was puzzled when setting x and y to 0 didn't cause
one sprite with the same origin as another sprite to become a copy.
Thinking about it, this behavior does make sense on the basis that
the projection methods use the sprite's origin as well, while all
the other draw methods draw based on the top-left corner of the
sprite being drawn to. So it's the same, but I clarified it a bit
in the documentation.