Reduced the delay between music ending and the next queued music starting.
There is always going to be a delay between these two, and this
delay is always going to increase as the frame rate decreases. But
I noticed that the previous method had a delay of not just one frame,
but two frames. This made the delay noticeable at 30 FPS. By
changing the method to use a Pygame music end event, the delay has
been reduced down to one frame, which is not noticeable at 30 FPS.