Reset the window size when exiting fullscreen.
this little bug was annoying: whenever fullscreen mode was exited,
the window would still be the same size as it was in fullscreen.
Now, it reverts  the window to an unscaled size if dynamic scaling
is on.

Also, when dynamic scaling is off, resizing the window (in windowed
mode) is no longer allowed.