Removed the default connection of event_room_start to event_room_resume.
sge.Room.event_room_resume previously called sge.Room.event_room_start
by default, the idea being that this would let people just define things
that need to be done every time the room starts or resumes there.
However, this is silly; it's easy enough to add a call to
event_room_resume in event_room_start and achieve the same effect,
and it meant that if you wanted a start event to only do things when
the room actually started, you had to define a blank resume event.