Reworked the collision detection system a bit.
"detects_collisions" was a bit vague, so I replaced it with two more
specific variables: "checks_collisions", and "tangible". The old
"detects_collisions" variable did an undefined combination of the
two.

"checks_collisions" controls whether or not the given object performs
collision checks automatically for the purpose of collision events.
This does not exclude the object from collision detection by other
means (e.g. collision checks by other objects, or direct calls to
the collision method). This is useful for objects which are numerous,
don't need to collide with each other, and which only need to collide
with a smaller number of objects, such as tiles.

"tangible" controls whether or not the given object can be involved
in collision detection at all. This is closer to what the old
detects_collisions was supposed to do. Since collisions can't be
detected, "checks_collisions" is always treated as False if "tangible"
is False.

This more well-defined system is now demonstrated in the examples.
Because of their simplicity, they don't gain much improvement, but
bigger games are improved a lot more noticeably.