Added sge.Game.collision_events_enabled.
I found that I could save even more frame time by disabling the
collision event system entirely, and this is a lot easier to do
than setting checks_collisions on every single object individually
anyway. So it is now possible to disable collision events entirely.

I am not going to be adding the ability to disable other kinds of
events. This can't bring any noticeable performance benefits, and
it isn't a shorthand for anything else.