Fixed a long-standing bug with iteration.
This bug was introduced sometime between 0.7.0 and 0.8.0. In version
0.7.0, the StellarClass.destroy method basically built a new list
without the destroyed object. Sometime in the development of 0.8.0,
that changed when I learned of the list.remove method, which was
used instead. Unfortunately, this causes problems when the room's
object list is being iterated through when the object was destroyed;
this manifested itself in the circle popper example by causing pairs
of circles to sometimes not both be destroyed when clicked at the
same time, since the mouse click event iterated through this list.