Added new stuff to make custom loops possible.
It occurred to me that there was no easy way to make custom loops
in the SGE when I noticed that I would need to do this for modal
dialog boxes in xsge.gui. A huge number of changes have been made
to fix this. Basically, the basic components needed for custom
loops have been moved out of the main loop and into separate
methods. An event system similar to Pygame's has also been added
in the form of sge.Game.input_events and the classes in
sge.input.

sge.Game.pause in the Pygame SGE has also been changed to take
advantage of this new feature. Incidentally, the new method it uses
is much better than the old one.