Added draw_lock and draw_unlock to sge.Sprite.
It occurred to me that drawing several times in a row on a sprite
is horribly inefficient, since the Pygame SGE has to do scaling
every time it happens to make sure it shows up on the screen right,
and it also occurred to me that such drawing can be useful sometimes,
e.g. for tiling. In light of this, I added the draw_lock and
draw_unlock methods to sge.Sprite, allowing multiple drawings in a
row to be made no less efficient than one drawing.

This is taken advantage of in the matrix wipe transition of
xsge.transitions. On my laptop, this transition previously slowed
down the game to 15 FPS. Now, it goes at approximately the full
60 FPS.