Fixed a crash when too much text was entered.
I had to think for a bit about whether this was a Pygame SGE bug
or an xsge.gui bug, but I decided it's an xsge.gui bug; it's
unreasonable to expect the SGE to automatically know that a draw
operation you told it to do is unnecessary.  xsge.gui now only
projects text that is currently visible in a textbox (and a small
amount of extra text on either side).

It slows down to a crawl at 30,000 characters or so, but that's to
be expected and not really a problem. I might add a character limit
to these boxes in the future so no one accidentally puts too much
text into these.