Added sge.View.wport, sge.View.hport to the spec.
These new attributes are designed to facilitate zoom support. This
feature exists in Game Maker, though at least in older versions it
was buggy and its use was not recommended because of this. This
will not be the case for the SGE; zoom support via mismatching view
size and view port size will be considered a perfectly normal
practice.