Removed class registration.
Class registration was originally added so that the SGE could
internally take advantage of it to make collision detection more
efficient. This idea has been superseded by a much better method.
Any use for the user can easily be replicated with lists on the
user's end, and even made better.