Changed the separate collision methods design to a single method.
While developing reTux, I noticed that, in trying to do the same
thing with two directions of collision detection, it was possible
for that thing to happen *twice* if both directions of collision
were detected. To fix this, I'm unifying all the collision methods
into event_collision, with new xdirection and ydirection arguments.