Fixed objects detecting themselves as a collisions.
A recent change to sge.Object.collision seems to have removed a
check making sure ``self`` isn't included in the results. This has
been fixed.

Funny enough, this bug wasn't demonstrated in either of my games.
This is because any time two objects of the same class collided,
it was handled by collision events, the mechanism of which has its
own prevention of self-collisions implemented.