Replaced code that converted a path to an absolute path.
The conversion was there to avoid errors caused by implicit relative
paths. Unfortunately, Pygame seems to be unable to handle any path
strings that have non-ASCII characters in them, so this was able to
cause an error gratuitously. To minimize this problem, empty values
are now converted to ``os.curdir`` instead.

It should be emphasized that this does *not* fix the problem. It
only prevents it from propogating where it otherwise wouldn't have.
The core of the problem is in Pygame.