Fixed problems caused by starting new rooms immediately.
This was a very subtle problem, and didn't usually cause noticeable
effects, but when it did have an effect, it could be catastrophic.
Because the effect of starting a new room happened immediately, the
remainder of the loop continued with the current room technically not
being the room the objects being manipulated were in. This was able
to cause all kinds of havoc.

I finally discovered this in ReTux when one circumstance actually
caused the game to crash, because code executed made an assumption
(which should have always been correct) that the current room had a
certain attribute. However, it seems this was also the cause of the
strange effect in ReTux of causing objects to fall through floors
when coming back to previous rooms: some objects were not yet
handled, and when they were, they were not in the list of "colliders"
due to sge.Room.start's action, and that caused them to start
falling -- which, of course, led to them falling through the floors
they were standing on.

So now, rather than starting the room immediately, sge.Room.start
instead schedules the room to start at the beginning of the next
frame.