Made rotation clockwise instead of counter-clockwise.
While rotation on the unit circle is counter-clockwise, it makes
more sense, mathematically, for rotation to be clockwise in the SGE.
The reason for this is the Y axis is inverted. By making rotation
clockwise, functions like sin and atan2 can be used directly without
first inverting the Y value.