Added some more Pygame bugs to the "known problems" section.
I was confused about what was causing a problem with the font sprite
in Pacewar, and it seems the source of the problem is actually
Pygame. For some reason, Pygame isn't giving the loaded image the
SRCALPHA flag, even though it does have alpha transparency.

I might be able to work around this by doing a check of all pixels
after loading the image and seeing if any of them have alpha values
less than 255, calling convert_alpha() on them if that is the case.
(convert_alpha() fixes the problem.)