Fixed the transparency problem.
It seems the problem was actually that Pygame actually loads some
kinds of transparency, notably the kind used by 8-bit images, as a
colorkey, because that's technically what it is. I was assuming that
all images with transparency would be treated as having alpha
transparency. Now that I know better, the bug has been fixed.