Added sge.s.
This is based on a suggestion by someone on the FreeGameDev.net forum.
He suggested that the use of strings for things like keys could cause
invisible bugs. So sge.s simply has several variables for strings
of particular importance in the SGE, so that someone concerned about
the potential for such bugs can use these variables; a typo will
result in an error.