Fixed a design that delayed garbage collection of sprites.
The cache included a strong reference to any sprite whose images
were cached and haven't been pruned yet. This is senseless, as the
reference only served to confirm what sprite the cached image was
for. To correct this, the cache now uses weak references for this
purpose. This means that while cached images will still not be
pruned until 30 seconds after their last use, garbage collection of
the sprites themselves will not be delayed because of this.