Fixed a bug in how the size of spritefont text was determined.
Well, two, actually. The first one was incorrectly accounting for
hsep and vsep (vsep shouldn't be accounted for at all, while hsep
shouldn't be accounted for as much). The second one is one I don't
quite understand, but for some reason, an extra pixel of width
and height needs to be added, or else the right and bottom of the
rendered text gets cut off. At least, that happens with some fonts.

Text seems to render correctly now, although it may come at a cost
of shifting text to the left by a pixel in some cases.