Removed support for implicit colorkey transparency.
I was getting hints that this was a bad idea already, but the last
straw was a bug resulting from lacking a way to detect transparency
in some images (these seem to be images that have transparent and
black pixels at the same time). Better to just require either alpha
transparency or an *explicit* colorkey. This will also prevent
changes in how a game looks from recompression of PNG images from
tools such as Trimage which should be completely harmless.

This change does not break compatibility with existing games, but it
does mean that some games may have new graphical errors (ReTux does,
for example), so the minor version number is being bumped. Games
can return to the way they were either by changing the images to use
alpha transparency or by specifying the colorkey explicitly in the
source code.