Added code to finally suppress that warning message.
It turns out that the warning was not a sign of a problem, but
due to some redundant code execution that came too early. The
warning has been suppressed by preventing the early and unnecessary
execution of this code. (For the record, it was completely
harmless and this results in no practical change.)