Fixed a bug where the collision functions would include intangible objects.
Man, these functions were so broken it's not even funny. I guess that's
what I get for not testing them...

This was found in Hexoshi, where the rectangle collision function is
used to determine whether or not the player can stand up. Opened doors
were counting toward this check even though they are not tangible,
causing opened doors to lock you into place.