Added support for sounds having "parent" sounds.
I was thinking for a while about how to make it so that two sounds
share channels, and this is what I came up with. When a sound has a
"parent" sound, the two sounds share the same channels (controlled
by the parent), allowing you to make these sounds conflict with each
other.