Modified how fullscreen is handled to make it more reliable.
It turns out that Pygame's detection of what resolutions are supported
is flawed. On Windows, some systems may appear to support higher
resolutions than they really do. On X, if the user is in a resolution
lower than native, X will crash the game when you attempt to use the
native resolution.

Using the current resolution is a better practice, anyway.