Added an optimization for games which are not currently being scaled.
r.game_display_surface represents the unscaled display (what the
developer works with), but if the window is *exactly* the same size
as it (meaning there is no scaling and no letterbox/pillarbox), then
it can simply be bypassed in favor of blitting directly to
r.game_window. Like the optimization in the last commit, this is
small, but measurable. In ReTux, it's a couple extra frames per
second.