Fixed a problem that could cause loss of frames in some cases.
Basically, when changing the animation speed, you still keep whatever
time already passed. That's intentional. However, when you do this
drastically and also change image_index, perhaps because you're using
a different sprite, you expect it to start from the beginning, but
it wouldn't always do so if you were increasing animation speed;
frames could be skipped.

To address this, I have made it so that the timing is reset any time
image_index or sprite is manually changed. That way, timing only
carries over when it's supposed to.