Fixed a bug in the music queuing.
This stemmed from a misunderstanding of how the Pygame music end event
works. Pygame's documentation didn't mention that stopping the music
posts and end event. What this essentially meant was that if you stopped
the music, cleared the queue, and started a new music queue, the first
track in the queue would be skipped because of the end of the previous
music that you had stopped b efore the queue was even created. Fixed
this by creating a "music end block event" which is posted when music is
stopped while no queue exists, or when the queue is cleared while a music
end event is posted.