Raise TypeError when using an invalid type in gfx.Color.
This also adjusts the implementation of how integers are parsed.
The previous method had the exact same results, but the new method
is more concise and makes it clearer what its intentions are.

Just for fun, I'll mention that I almost changed it to this:

(value & (1 << 8 * (i+1)) - 1) >> 8*i

...which is basically just a rewrite of the original expression
showing what it did more clearly. Um, yeah, bit-shifting first and
then just &'ing a static 255 is much nicer.