Added hexagonal tile rendering.
Side note, Tiled really really ought to document this stuff. I spent
hours trying to figure out what to do here, assuming incorrectly that
math needed to be used somehow, only to notice that, actually, this
is completely arbitrary. If Tiled's documentation mentioned that each
row is simply rendered exactly 75% of its height below the position of
the previous row, it would have saved me a lot of trouble. 😑