Added a simple shader API.
It's not very efficient. I really really tried to get this done
more efficiently with Pygame's pixelarray API, but the thing segfaulted
in my tests and I for the life of me couldn't figure out how to actually
iterate through a pixel array properly, or to convert a 2-D array to
a 1-D array (which seems to be necessary since apparently Pygame gives
you a 1-D or 2-D array depending on how it feels). And surfarray relies
on instantaneous garbage collection, so that's out. But whatever, I
guess. I don't even know if there would be an actual performance benefit;
that would depend on how well get_at and set_at are implemented.