Fixed attempting to use Pygame colors, added example to Circle Popper.
Having tested this now, the shader system is tremendously slow. Circle
Popper previously ran at 900 FPS, but with just a basic shader on a
tiny 32x32 area, it slows down to around 150 FPS. We'll have to find
a more efficient way to do this or otherwise figure out how to make
these shaders work efficiently.