Changed ordering of setup steps for adding an object to a room.
This moves the create event to *after* other setup steps, preventing
warnings and a small potential (albeit unlikely) for issues. I'm
actually not sure why it was set to do the create event first, but I
didn't encounter any issues from changing this arising in either
ReTux or Hexoshi (and if there's any problem with the seemingly logical
ordering then I would think one of those games to be most likely to
show it).