Massively improve efficiency of Sprite.swap_color.
This switches sge.gfx.Sprite.swap_color to using a method more
similar to shaders. The attempt to use real palette swapping has been
discarded because all sprites in the SGE are converted to use alpha
transparency, meaning they will always be in value mode rather than
indexed mode. As such, I've also removed the suggestion to use indexed
mode graphics, and replaced it with a note that use of draw_shader()
is likely to be more efficient if you're replacing multiple colors.