Removed use of deprecated Pygame APIs which seem to be broken.
Apparently, on Windows, the pygame.ACTIVEEVENT event type is horribly
broken on Windows, because it doesn't have the mandatory "gain" attribute.
This has to be the worst possible way the Pygame project could deprecate
an API, but given that, it's clear we cannot safely continue to use
deprecated or "legacy" APIs. As such, I've updated use of ACTIVEEVENT
to the new event types, and I've also updated the joystick events to
use the new "instance_id" attribute instead of the deprecated "joy"
attribute.