Made a change which might fix the problem with the pixelate transition.
Digging thrū the code, I noticed that the function that handles the
pixelate transition accidentally calls pygame.display.set_mode(),
related to a hack that hotswaps between the actual scale mode and the
default scale mode. Made a change which should work around that by
doing the hotswap with internal values so that r._set_mode never gets
called.